Fix split editor shortcuts not working if in Design mode
When using the split editor keyboard shortcuts to navigate to "Design",
we were setting the focus to the split editor component only if the
delegate was equal to SplitEditor.getShowPreviewAction(). This was not
true for some SplitEditor subclasses that override this method, such as
DesignToolsSplitEditor, consequently making the editor to be "stuck" in
design mode when navigating to it using the shortcuts. This CL fixes this
issue by requesting the focus to the editor component if the current action
is equal to showPreviewAction, which will work for all subclasses.

While fixing the issue, another problem was exposed: changing the focus
to editor.component would not work, since EditorImpl removes "myComponent"
from the UI hierarchy manually. In order to properly give the focus to the
editor, we should request the focus to editor.contentComponent.

Fixes: 147739570
Test: Use the keyboard shortcut to navigate modes in split editor.
Change-Id: I75e58308d72b1930039391f6f349bd14dc1af33c