Do not use a separate thread for RenderTask dispose
Having a separate thread can cause leaks since the thread had a 10s
timeout and not way to explicitely dispose it. With this change we just
move the dispose actions to be run on the common pool on a
BoundedExecutor. Since the common pool is already disposed by IntelliJ,
the thread will not be leaked.

Fixes: 127817242
Test: Existing tests pass
Change-Id: Idb383946036994a69ef1fdc872a0430f40cb47b6