[Dsl,.kts] parse some more method calls as blocks
In KotlinScript, we already have some method calls parsed as blocks:
specifically, those corresponding to elements of
NamedDomainObjectContainers, which have to be looked up or created by
name using operators such as getByName("foo") or create("foo").

Analogous to the Groovy case, in order to better support repositories
being declared using simple method calls, we need to parse those as
blocks too, whether or not they have closure arguments.  In order to
do that, isBlockElement() needs to be passed the parent element, so
that it can determine whether the given method call is one that should
be interpreted as a block given that context.  The same special cases
apply as in the Groovy case, with the additional wrinkle that bare
configuration specifiers (e.g. configurations { create("foo") } ) are
also method calls which must be handled as blocks -- the Groovy
equivalent is configurations { foo } which is not parsed as any kind
of method call.

Bug: 144679123
Test: existing, further tests coming
Change-Id: I94d17c8f45d7972cc5378199a7a86ad824bd5fc3