[Dsl] start handling repositories properties
Now that the parsers are willing to construct block dsl elements for
the bare method calls in repositories blocks, we can integrate those
repository dsl elements into the property renaming mechanisms recently
developed and integrated.

We special-case MavenRepositoryDslElement in the writer, because empty
"blocks" of that class should be written out as 0-argument method
calls.  This differs from other empty blocks, such as bare references
for configurations or buildType / productFlavor.

Now that we never have method calls to represent repositories, we can
adjust constructors for the RepositoryModel concrete classes to always
accept a GradleDslPropertiesElement for the corresponding Dsl element,
allowing us to simplify the model's `name` property.  Handling the url
property is a little more involved than this (changes forthcoming),
because of the need for a language-specific single-argument method
transform to handle the need in KotlinScript for a uri() helper
function.

Bug: 141842964
Bug: 141970574
Bug: 144679123
Test: run RepositoriesModelTest.setNameForMethodCall for .kts
Change-Id: I6efb3e95240abd2e398e1c4e85aa97d1496f8dac