Old AGP shouldn't offer hyperlink to download NDK
In this bug an old AGP is used (3.1). It doesn't support the
property android.ndkVersion in build.gradle. In this case we
shouldn't offer a hyperlink to download latest NDK because
it will be installed at a location (ndk\<version>) that will
not be recognized by AGP (expected is ndk-bundle).

Bug: 147832066
Test: Added unittest
Change-Id: I26b6ce0fa85471fb290fe74467e4cafde2b5f845