Refactor the model for enum values for properties
Before EnumValue held values for a header before the value and a
separator after the value. This was done such that the user was
unable to select a header and separator in a list.

After this change headers and separators are treated as separate
enum values. These new enum values are marked with a Selectable
interface that gives a clue to the UI that these values should
never be selected.
Notice that the horizontal margins now is added in:
DarculaComboBoxUI.customizeListRendererComponent where before
the margin was set in:
EnumValueListCellRenderer.customizeCellRenderer.

The bug 145551681 was caused by the previous renderer that would
create a panel with the header & item (and optional a separator).
A change in Intellij caused added padding to the resulting panel.

A selected element would consist only of the text portion not the
header and separator thus the selection did not expand to the
side of the popup (see the bug).
We could have given the outer panel the color of the selection,
but then the headers and separators would appear inside the
selection.

To fix that problem I now treat headers and separators separate
from the elements in the list. This is probably a better design
as well.

Fixes: 145551681
Test: Existing tests should pass.

Change-Id: Ic4208868925117ce47a2636bdb9d190817e7ed3b