[Dsl,.kts] fix writing of new module dependencies
The implementation of addModule() for the Dsl dependencies model used
the configuration name of the dependency too many times: once for the
method call, and once for the map representing the method arguments.
The Groovy writer implementation ignored the latter name, but the
KotlinScript one did not, resulting in incorrect output (that our
existing tests were not sufficiently robust to catch).  Fix the
double-use of name and strengthen the tests.

Bug: 142114586
Test: existing tests strengthened
Change-Id: Iac9b25c4d7b8e55a585cff393ad1886219dcb065
(cherry picked from commit d3a3f7b4831fc1e6b2dcd4000085a8900987b963)