Create correct type of android project view nodes
regardless of the presence of the AndroidModuleModel. The type of
a module is determined by its facets and does not change when sync is
running, fails or is just about to start.

This change eliminates long flickering of the Android project view in
the case when the previously selected node is in a source code folder
of the main source set of a module. The top level nodes are still
reloaded but the processes is not repeated for all node in the tree.

The change does not solve visually the same problem in some other cases.

Bug: 111527585
Test: n/a (existing project view snapshots)
Change-Id: Icc2dc35f5be7ff9aff7599efe848c93be91a38af
(cherry picked from commit aa6ad48158365fb429c285a32910b8bf1c15dbc0)