Register ComposePreviewRunConfigurationType in XML
Register ComposePreviewRunConfigurationType in compose-designer.xml, and
control its activation through the COMPOSE_PREVIEW_RUN_CONFIGURATION flag.
If the flag is disabled, we throw a ExtensionNotApplicableException and
that will guarantee the Compose Preview run configuration is not
displayed in the UI.

The change described above, by itself, would cause when IntelliJ APIs
were used to iterate through the list of RunConfigurationProducers.
ComposePreviewRunConfigurationProducer would try to return an instance of
ComposePreviewRunConfigurationType and ExtensionNotApplicableException
would be thrown. To fix that, we created a StartupActivity to add the
producer to a list of ignored producers. This way the producer's
getConfigurationFactory won't be called and we can actually simplify it,
returning the proper instance of ComposePreviewRunConfigurationType.

Bug: 144500928
Test: ComposePreviewRunConfigurationTest
Change-Id: Ife049be84cf7294e758f2dc019ed238c605abdb9