[Dsl] improve buildscript / variable resolution
An expression in buildscript { ... } context will be executed before
the remainder of the build file for that module.  This means that if
we are resolving a variable in such an expression, we must not
consider any of the toplevel variables.  This was already implemented
for the simple form of referring to extra properties (as simple
variable references) in Groovy, but KotlinScript does not have such a
simple syntax for extra properties defined without delegation.

When resolving an expression in buildscript { ... } context,
references to the current project (such as through the `project`
property) are implicitly converted to the Buildscript property of that
project, if any.  Other projects (e.g. parent and rootProject) are
already fully configured and both regular and buildscript content are
therefore available.

The test for writing buildscript ext properties did not assert the
form of the output: which is important given our somewhat lax parsing:
a successful parse does not necessarily imply a successful write.
Check the exact form of the output, for both Groovy and KotlinScript.

bug: 142114586
Bug: 147796130
Bug: 148271448
Test: existing tests strengthened, run for KotlinScript
Change-Id: I8108d737eafced0d60ebcec43dda513c2e7af709