Use a fake NDK package for testing.
IdeSdksTest includes a couple of tests that verify that the NDK has been
set up properly as part of setting up the Android SDK. These tests don't
really need anything from the NDK - they just need the NDK package to be
available in the SDK Manager.

Currently, they use the real NDK, which necessitates a long set up time
for this test as over 10k files in the NDK need to be made available to
this test. This CL fixes the issue by providing a fake NDK to the SDK
repository loader.

Bug: 148236162
Test: N/A
Change-Id: I8f108878b379c5530b077fdda52c565c130f9fb1