Add empty views to AndroidTestSuiteDetailsView as a placeholder.
DetailsViewDeviceSelectorListView will take place at left side of
the TestSuiteDetailsView, where you can choose a device to look up
test results. DetailsViewContentView will take place right side of
the TestSuiteDetailsView. The test results of the selected device will
be displayed in the DetailsViewContentView.

See the buganizer for the spec and design mocks.

Bug: 143299082
Test: unit tests
Change-Id: If310e1752ac16409fadbf7a70d504f87bea645c8