[Dsl,.kts] quote unknown configurations when renaming
In KotlinScript, dependency configurations that don't have specific
extension functions must be specified with strings, themselves handled
with a catch-all extension function.  This must happen when setting an
existing dependency's configuration (for example through the Dsl
model's `setConfigurationName()`) as well as when creating a new
dependency in such a configuration.  Handle the former case by special
treatment in KotlinDslWriter's maybeUpdateName().

The current implementation of renaming in KotlinScript will preserve
the string-quoted nature of existing dependency configurations, even
if they are being renamed to a known configuration.  This might be
slightly suboptimal aesthetically but does not break the project.

Bug: 138277630
Test: added, existing tests fixed
Change-Id: I6e4b063005c64ebdc88f3c0e2ebf384c38c3c0ac