[Dsl,.kts] set map properties using setter methods
Changes likely to see release in AGP 4.1 (or whatever version it gets)
alter the way that assignment to map properties (manifestPlaceholders
and testInstrumentationRunnerArguments) is handled: the conversion of
the API to Kotlin means that mutability is explicit.

In order that build configuration emitted by Studio 4.0 remain
compatible with the new way of doing things, make the Dsl prefer
writing build configuration of these map properties using setter
methods.  (This happened to reveal an uncovered case in the Kotlin Dsl
writer, of emitting a map property with an explicit name).

Bug: 140406102
Bug: 148657110
Test: updated and re-enabled tests
Change-Id: I7212aeaeb516ef11352451b266109795170edf2d