Add the current text to the auto complete.
There were ambiguities what <enter> would do when completions were
shown in an editor in the properties panel. See the bug.

We also attempted to move the best match to the top before this CL.
This was removed since it didn't quite work as intended and now we
are showing the current text at the top spot in the completions
instead.

Fixes: 148628592
Test: Existing.
Change-Id: I0c3535f93268b0cca90693756685a2c45c8c604e