[Dsl] support names which match language keywords
NamedDomainObjectContainers can contain elements named by (almost)
arbitrary strings: including strings which also happen to name
keywords in whatever build configuration language is being used.

KotlinScript's lookup is by string anyway, which means that this case
is already handled; Groovy must quote the name if it matches a
language keyword.

(This does not solve the more generic problem for variable names,
which is problematic in Groovy as quoted identifiers are not supported
as variable names, but it does further mitigate the potential problems
coming from user input at the PSD's build variants configurable)

Bug: 126937269
Test: added
Change-Id: I4e4accce9f865436e33d7055414b8f479a3207e2