Change warning message for JDK location
... to suggest that the path might be different, instead of stating they
are. Also, there have been some confussion from users about how the
paths reported in the warning message are used.

With the ne message, we no longer compare what Android Studio uses to
what a terminal might be using, since we cannot tell for sure if that
would be the case.

Test: existing test should pass, changed a warning message only.
Bug: 143402226
Change-Id: I6165b3d8355f6c3da62a6a9feffa998c08a6bcf2