Add the newly create property instance to the model.
When a new custom attribute is created, we used to regenerate all
attributes for the "new" selection. Now that we are smarter and
only invalidate the current values we need to add the newly
generated property to the model.

Fixes: 148655617
Test: Unit test added.
Change-Id: Id234e75b2569138551d4f0cb57132ae37e0a7d62
(cherry picked from commit e237420b3f2a84fd0a5d6981a953f8aed43c0185)