Don't start profiler session before events arrive
There's a bug when the user stops the profiler session and
then clicks the Run button. A explanation of the bug is
documented in the comment in StudioProfilers.java.

Because of the fix, also tighten the logic in the processing
order of session updates.

Bug: 148519288
Test: updated & manually, mixing Run and Profile button and two devices
Change-Id: Ie8282ae25ad198a17e5c027a0fd8beafa823404b
(cherry picked from commit 78622dabda07889de3b67cd1cc74666482b77d04)