Move painting responsability down to the SceneView
SceneView own the Layers so there is no need for the surface to handle
the painting of the SceneView Layers. This CL refactors the SceneView
and the surfaces to remove Layer painting from the surface and moving it
to the SceneView.

This will allow in the future to simplify the SceneView layout and
painting so we can easily add the top toolbar.

The CL also changes the API to allow LayoutManagers from relying on the
list of SceneViews to be sorted by any order. It is the responsability
of the LayoutManager to sort them in any way they want.

Test: Existing tests pass
Bug: 148859613
Change-Id: I7c0ce8d8fd2f6e5520c352c19f6ace27f1cc199a