[Res Explorer] Have the ChessBoardPanel take a cellSize argument
Make the ChessBoardPanel a bit more flexible, so that it can render the
Chess Board at a different size/scale.

This is so that the ChessBoardPanel can be used at lower dimensions,
otherwise the chess board pattern might look too big.

As a side effect, this also makes it so that it scales properly when
changing in and out of presentation mode.

Bug: 148391228
Test: manual
Change-Id: Ia9e5feafb07fd5ad3ed71f7cd73e72e779f0f238