Fixes around switching devices
- Previously we were not correctly filtering on stream and pid, since
selectedStream and selectedProcess were the defaults at the time the
listeners were registered, and those objects were always used to get the
filter values.
- Previously we weren't detaching and stopping the agent when switching
selected processes.
- Previously we were not resetting the last seen event timestamp here or
in TransportEventPoller (done implicitly by removing the listeners) when
switching devices. This meant that if one device had a timestamp
significantly ahead of another, all events from the device with the
earlier timestamp would be filtered out after switching to it.

Fixes: 147491720
Bug: 144058415
Test: None for now. Hopefully this will all be changed when we switch to
Test: the app inspection API.
Change-Id: I99cbc9df426a1a0ba4fb3e234fcfb1a273168073