Add TransportStreamManager and TransportStreamChannel
When we encountered the issue of out-of-sync timestamps across different
streams, it was clear that StreamId need to be enforced when listening
to all events except for streams. It would be convenient to have a
a central stream "manager" that automatically hands new streams
encapsulated as stream "channels" to interested parties. These channels
can then be used to register event listeners that are automatically
associated with the underlying stream.

This CL introduces TransportStreamManager and TransportStreamChannel.
TransportStreamManager listens for and creates new
TransportStreamChannel for every single new stream. It has a listener
interface thru which it hands TransportStreamChannel to listeners.

This CL migrates app inspection framework code to use the new
TransportStreamChannel away from TransportEventPoller.

Note: this CL does not try to fix TransportEventPoller.

Test: Updated tests. Added TransportStreamManagerTest
Bug: 148933826
Change-Id: I562e4503de7fc72e24f8f8341d93710d445bd7e8