[Dsl] probably make moduleWithDirectory linear
Since we generate the list of standardized module paths from the Dsl
by calling modulePaths(), we don't then have to check whether the
module path we have passed to moduleDirectory() is a valid one.

Since we almost certainly call moduleWithDirectory once per module in
the system, we probably still have a scaling problem.  It should be
straightforward to make it amortized-constant time if necessary.

Bug: 149029879
Test: not yet
Change-Id: Iedf28656da9b4eb2d20cbd4ddd1d1d11f768d0c4