Fix project root detection in the variant selector.
Selected build variants are stored in the project's facets and are
available even when Gradle model cache is not available, for example
early in the process of opening a project. Therefore, if the variant
selector relies on any Gradle models to restore the selection of
variants, the selection is lost if sync is launched while opening
the project.

This change fixes the problem with non-native variant selection being
lost when opening a project that requires sync. The same problem for
native variants remains unresolved.

Bug: 142608498
Test: testSwitchingVariantsWithReopenAndResync_simpleApplication enabled
Change-Id: I96e3840515e6721188d1a986f7f8c4492097d4dc