[PSD] allow setting dependenciesInfo.includeInApk
Conditionalizing when to display the property combo box does not quite
work as expected: the configurables are not refreshed when the plugin
version (or indeed anything else) is changed, so a change from
e.g. 3.5.3 to 4.0.0-dev would need the user to exit the PSD and
re-open it to see the includeInApk combo box in the Modules
configurable.

Bug: 145728802
Test: manual
Change-Id: If9e7163b4381f26870453cb6dffd607041b9f7c9