149203366 Fix Crash report: AssertionError at LintIdeClient.report-39094cea
The root cause was that there are cases where for example the
layout editor reuses lint infrastructure (in particular, ApiLookup)
which requires for example a LintClient and a LintResult. However,
the default LintResult class does not expect to receive error messages,
which can happen if for example the ApiLookup cannot find the
configured SDK Home (perhaps because the user has not configured it).

This CL makes this more explicit. Instead of just removing the
assertion, I switched the result classes to a sealed class hierarchy
and created an explicit ignored result class.

Test: Existing
Fixes: 149203366
Change-Id: Ib64f47869e440196a4758d0c8fedbc45f9405d74
(cherry picked from commit e90695d7da4f6aec200c21459f221d8a62c373cf)