Update the actions toolbar on model changes.
Previously the actions toolbar would not update if the status of the
model changed. The toolbar update would be triggered by an unrelated
mouse move by the user. This toolbar update was surprising for the
user and a mouse move in a ui test did not provide a consistent update
of the toolbar.

In this CL a model update will cause (a potential delayed) update of
the actions toolbar.

Bug: 145520901
Test: Will be tested in a followup CL which adds a ui test.
Change-Id: I59ad04a86c32cfcb6251cbd89891d8410cc7c0b7