Get build output listing file from VariantBuildInformation
VariantBuildInformation was added to AGP to provide
minimal information for variants, such as build task names,
output listing file location and etc. It is designed to replace
the listing file fields in AndroidArtifact model, and to provide
task names for un-synced variants.

This change makes use of the new VariantBuildInformation model
to get output listing files for non-test variants.
We have to rely on the old fields in AndroidArtifact for test
variants because VariantBuildInformation doesn't contain test
artifacts at this time.

Move all related methods to a utility class GradleBuildOutputUtil.

Bug: 146352662
Test: Covered by GradleApkProviderTest
Change-Id: Iebbd63315fd0e960df3e174c67d8c850579ea2a5