[Res Picker] Allow to provide a context when rendering drawables
For the Resource Picker, when rendering drawables, we want to render
them in the same theme as the layout that they'll be added to, so we now
allow to pass a contextFile to DrawableAssetRenderer.

When no context file is provided the drawable simply will be rendered
with its own set theme, which is the case with the Resource Manager.

Since both renderers now have to create a render task for every drawable
render request, rendering might be considerably slow, however, this also
saves us the trouble to worry about leaking the render task, not to
mention this is done to guarantee that Drawables previews are as correct
as possible.

Deleted unused class DesignAssetDetailViewModel.

Bug: 145138478
Test: Updated
Change-Id: I7439ace0f063f11a679742737ab97e3aaf473778