Unregister the attach listener by using return value.
This fixes a leak in the layout inspector:
We used to unregister the attach listener as part of the callback.
Unfortunately the poller would store the last timestamp after the
callback ended i.e. after the listener should have been
unregistered. The poller actually had code to deal with this by
simply returning true to make the poller unregister the listener
after the callback was called.

Bug: 145520901
Test: This will be covered in a followup CL where a Bleak test is
      added to the new layout inspector.

Change-Id: Ic170c4235ef0925eed9d39f29ef7e0c85405ca14