[Dsl] handle JavaVersion.VERSION_X_Y
In order to not treat JavaVersion.VERSION_X_Y as an unresolved
reference, we have to resolve it to something, which means that it
will need a value.  But what value?  Unfortunately, the various
possibilities (String, ...) all come with assumptions about their
externalisation, which are already used in the implementation of
LanguageLevelModel and its clients.

We therefore define a new kind of Dsl element, an Enum, to hold a
collection of enum values (interpreted as properties).  At the moment,
all we need it for is to be a container for textual values; we should
never be setting this kind of value, nor should we be really enquiring
about its true value.  Arrange for getValueType() on
LanguageLevelModel to return CUSTOM if the value is one of our enums,
and arrange for JavaVersion to be such a global enum with appropriate
properties in a GradleDslFile.

Bug: 149206270
Test: existing tests updated
Change-Id: I7549f709be537e7137938f9ffb1d1d851e8ab0d8