Focus on the editor preferred component after changing state
After changing state in the SplitEditor, focus on the preferred
component of the editor for that new state.
That way, switching to Design for non-editable previews will still focus
on the DesignSurface instead of the EditorComponent, allowing all the
shortcuts registered to the surface to work (including zoom).

Bug: 149816806
Test: manual check
Change-Id: Ifca7278ea9312e10154d6a7b44c9e9d174f933f6
(cherry picked from commit 39dc64d515b966c4d210675a21b6f3feb85c122f)