Add logic to report state of Build Analyzer
In this cl I define the states of the BA feature enum
and a manager responsible for providing current state to any
interested party as well as notifying the subscribers
about the changes.
Currently the state is about if we have the data to show already
or what is the reason we don't.
At this moment this information is required for implementing
the button in "What's new assistant" with the corresponding states.

Bug: 148672778
Test: existing updated and new added
Change-Id: Ibbea93adf3797ca2dad3034fed77335289988b6a
(cherry picked from commit bc11674d5c5c6f82698567d4edfde5562afa2f38)