[Gradle Sync] Extract method to convert Module to display name...
from the AndroidViewModuleNode to a new AndroidGradleUtil file. This
change creates a new file instead of reusing one of the existing util
classes as in the future it would be good to prune and merge all of the
current util classes into a few with improved documentations. This new
file will attempt to be the start of that.

This change also changes SyncIssue reporting to use this module name
over the name of the module in preperation for qualified module names.

Bug: N/A
Test: updated
Change-Id: I0cfa4c806fc850975b42db7a27bfa3ba0fc86147