[Res Picker] Create a Compact Resource Picker
Added CompactResourcePicker, functionally, it's just a substitute of the
ColorResourcePicker. Underneath, it shares resource fetching and
rendering logic with the Resource Manager, to guarantee that both
components provide consistent results.

So resources are rendered in the same way and it uses the same cache for
images (when rendering drawables or layouts).

Additionally, it can now show Theme Attributes colors and it can be used
for any other type of resource.

Bug: 148391228
Test: CompactResourcePickerTest
Change-Id: Ic4bf7ed4ce46152d14c25862b7db4e50d6b84db7