Recalculate scroll area after adding first destination
AddDestinationMenu was calling update instead of requestRender.
requestRender will zoom to fit after the first destination is added. On
top of that, the zoom to fit would not recalculate the scroll area if
the zoom level remained constant. This would not layout components
correctly if they needed to be centered on screen (always after zoom to
fit).

Test: Manually tested
Fixes: 150163354
Change-Id: Ifedc16e37d8c32ab382a2a8ccce9fb1066423482