Refactor layout code to not depend on SceneView
This CL creates an abstraction for positioning content in the
DesignSurface. Instead of directly depending on the SceneView, all the
layout code will now depend on a PositionableContent abstract class.
This allows replacing the SceneView layout code with code that positions
the SceneViewPeerPanel without modifying the code itself.

This should also allow for the code to be tested independently of the
SceneViews. This abstraction can be removed once the code depends only
on Swing layout.

Bug: N/A
Test: Existing tests pass
Change-Id: If015336954c8c49b0b8241323d1ba50d97de29c0