Add process managers to hold current processes in memory.
Currently the "Set Process" button in the layout inspector will
cause an RPC. That RPC is run on the UI thread. Moving it off the
UI thread would cause a different problem where the UI always will
show the list of process that were known last time the button was
pressed.

This CL will keep track of the current devices and processes such
that the dropdown can be presented immediately.

Bug: 150319501
Test: Unit tests added.
Change-Id: I0943fc5081484a11339ac12ca99401e7c04e8029