Add process managers to hold current processes in memory.
Currently the "Set Process" button in the layout inspector will
cause an RPC. That RPC is run on the UI thread. Moving it off the
UI thread would cause a different problem where the UI always will
show the list of process that were known last time the button was
pressed.

This CL will keep track of the current devices and processes such
that the dropdown can be presented immediately.

For 4.0 we don't have the services needed to do this. We are missing
the TransportStreamManager from the change I562e45. Instead this
change makes RPC on the UI thread with a timeout.

Bug: 150319501
Test: Unit tests were added in studio-master-dev.
Change-Id: I0943fc5081484a11339ac12ca99401e7c04e8029