Add tests for separate heap-dump UI
This CL clones the old tests for legacy heap-dump UI and adjusts
it for the new UI. It exposes more high-level states in the new UI
(e.g. is showing capture object, etc.) to enable equivalent assertions
from the legacy one without being too specific (e.g. second component
of the splitter being null).
It also finds and fixes a few null-pointer errors.

(This CL is a redo of ag/10518954 that was reverted by ag/10626553
due to a flaky test)

Bug: 151114950
Test: add new
Change-Id: Ic15dfe276ea01de115d5793cc5ae661f9223d216