Handle errors from on-device inspector.
An on-device inspector can send an error as a response
to a command (synchronous) or as an event (asynchronous).
When detected, synchronous errors are thrown as exceptions
so that they become part of the usal flow for errors:
they cause the futures to fail and are shown in the views.
Asynchronous errors are delivered to
DatabaseInspectorProjectService that takes care of showing
them.

Errors from the on-device inspector have three fields:
message, stackTrace and isRecoverable.
Is recoverable is an indicator of the type of error.
If an error is not recevorable it means that the
on-device inspector must be restarted. This is communicated
to the user by appending a string to the error message,
saying that the app must be re-launched.

Test: LiveSqliteResultSetTest.kt, LiveDatabaseConnectionTest.kt, DatabaseInspectorCLientTest.kt
Bug: 151004578
Change-Id: I38950bee7b7a1cd94b5afc6ee21f061eab9f83f8