Select table cells when showing popup.
This CL fixes a bug that prevented table cells from being selected
before showing the popup.

This slipped through the tests because fakeUi sends a popup
event only with the mouse clicked event. But in swing
a popup trigger can be on mouse clicked, pressed and released.
We were covering only the click event case.

Test: TableViewImplTest.kt
Bug: 150862643
Change-Id: I1df2865bed2b833dce3b7fb97afce8917f107e6a