Fix UI pause when help panel is opened.
There was a report that UI freezes for few seconds when help panel is
opened for the first time. This is mainly due to all the images, gifs
are being loaded at the same time.

This change makes sure the card that's not visible doesn't load images
nor gifs unnnecessarily.

Bug: 150114209
Test: Manually tested by checking how often redraw is triggered when
help panel is opened for the first time

Change-Id: I44ab282eece3e73826941d4b0e99419632164077