Remove table from TableController.
Since the schema can be altered using ALTER TABLE
statements, the user can add columns to a table.
Therefore it is wrong to store the table in the controller,
it might become stale.

An example of a problm caused by this is when trying to update
a cell. If the initial table has a rowid column, that column
will be used to run updates. If, without colosing the tab,
the user runs an ALTER TABLE and adds a column with the same
name as rowid, the update cell action won't behave as expected
becuase now rowid is an user-defined table, and not the
actual rowid column.

Test: DatabaseInspectorControllerTest.kt, TableControllerTest.kt
Bug: 150928547
Change-Id: Ib716ac4cdfb205f6c0e98f7f593754a7ab7150e7