[Res Explorer] Support preview for Navigation resources
Under NAVIGATION_PREVIEW flag (enabled by default).

Enables a graphical preview for Navigation resources, the preview may
have the layout preview of the start destination in the Navigation
graph.

That means, to have a proper preview, tools:layout has to be set in the
start destination tag in the Navigation file.

Bug: 147157808
Test: NavigationSlowPreviewProviderTest
Change-Id: I0e0d82fd6fbeca559374b8b3ef90ca1f66b36f72