[Dsl] annotate elements with their effect on the model
In order to begin coupling parser and writer to Gradle model using
structured properties rather than strings, annotate Dsl elements
directly at parse-time with their effect on the model (if any).  This
removes the need for "canonizing" GradleNameElements.

We continue to look up model properties by name for now, which means
that this implementation is not as clean as it could be; at the
moment, the name of an element (from the perspective of the model)
could be held in its GradleNameElement, or derived from the
ModelEffect's property, or (in unusual cases) the hierarchy from the
GradleNameElement and the name from the effect's property.  This
less-than-ideal implementation should go away once the model also
operates on the level of properties and effects rather than strings.

Bug: 141842964
Bug: 148657110
Test: existing
Change-Id: I8766c96bd9c6e47b88d92ddcc21e6bfec5f7e883