[Dsl] support property descriptions in model builders
Now that Dsl elements are annotated with their effect on the build
model, the missing piece is to associate individual dsl model objects
with the model properties that they are modelling (rather than their
names).  To start with, support ModelPropertyDescription arguments to
model builders.

Note that this doesn't change the models themselves: the builder
extracts the property's name and uses that to build the model, rather
than using the property itself.

The model property MATCHING_FALLBACKS will be our test case, as an
upcoming change requires it to be treated as a mutableList in Kotlin,
rather than just a List.  Convert MATCHING_FALLBACKS to a
ModelPropertyDescription.

Bug: 141842964
Bug: 148657110
Test: existing
Change-Id: Ibcf3cd2912a30a826481915c9d52fc8d16593526