[Dsl] thread model property descriptions through the model
If we're building a model from an existing element, use the element's
effect, if any, to look up the property, but if we're building a model
for a property with no existing element, store the property in the
model.  (Building models for string names continues to operate as
before.)

When we bind (or bindList or bindMap) and in so doing create a new Dsl
element, annotate that new element with a model effect if our model
has a property.  The effect that we create should not have either VAR
or SET semantics, but merely express that the property acquires the
value, allowing Dsl writer backends to choose how to externalise
that.  (In practice the backends are still operating on names, so this
is unexercised, but this way when moving to operating on explicit
model property and effect descriptions we can test for the "acquire
value" semantics).

Bug: 141842964
Bug: 148657110
Test: existing
Change-Id: I3c1c7cb8b7eecb4dd2abe200939c80e81121dd1e