Do a new full layout after scrolling in ScrollViewScrollHandler
When calling View.scrollTo in layoutlib native using hardware rendering, the direct parent of
the scrolled view gets marked as dirty, but no one else in the hierarchy. So when asked to
re-render, layoutlib looks at the root view and does not know that anything has changed and
performs the same render as before the scrolling.
Calling requestLayout invalidates the entire hierarchy, which means that layoutlib then renders
correctly handling the scrolling.

Bug: 150949145
Test: checked manually
Change-Id: I870e099ea45dc2babc60effd543fdd680e588622