Add UI support for data invalidation.
This CL adds UI support to handle the DatabasePossiblyChanged
event from the on-device inspector.
Each time an event is received the schema of all open databases is
updated and each open table (with live updates enabled) is updated.
Live updates can be disabled or enable for each table using a checkbox.

Test: DatabaseInspectorControlerTest.kt, SqliteEvaluatorControllerTest.kt, TableControllerTest.kt, TableViewImplTest.kt
Bug: 152216087
Change-Id: I40dbf75c017076d04df8a8a2b1724c17a4fc8a2a