[Dsl,.gr] write method calls using external name information
The Groovy writer wrote method calls out directly, without consulting
the external name information about whether the property in question
should use assignment or not.  Changing this to look up the
information, as in the KotlinScript writer, allows removal of some
more asMethod information in the Dsl model, where it does not belong.

Bug: 141970574
Test: existing
Change-Id: Ie3f0c629e8047f5a7cf8fb618f2c17cea31d1e2f